Move ConfigValues to a separate module (#10)
